# Hw10 optimization is done