--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -1,67 +1,68 @@
 package hw10programoptimization
 
 import (
-	"encoding/json"
+	"bufio"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"regexp"
 	"strings"
+
+	json "github.com/mailru/easyjson"
 )
 
+// easyjson:json
 type User struct {
-	ID       int
-	Name     string
-	Username string
-	Email    string
-	Phone    string
-	Password string
-	Address  string
+	Email []byte `json:"email"`
 }
 
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
+	return getUsersAndCountDomainEntries(r, domain)
 }
 
-type users [100_000]User
-
-func getUsers(r io.Reader) (result users, err error) {
-	content, err := ioutil.ReadAll(r)
-	if err != nil {
-		return
-	}
-
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
-		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
-			return
-		}
-		result[i] = user
-	}
-	return
-}
-
-func countDomains(u users, domain string) (DomainStat, error) {
+func getUsersAndCountDomainEntries(reader io.Reader, domainName string) (DomainStat, error) {
+	user := &User{}
 	result := make(DomainStat)
+	var found [][]byte
+	var err error
 
-	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
-		if err != nil {
-			return nil, err
+	// usage of methods from strings package actually won't save much more cpu, than precompiled regex on byte stream,
+	// but will consume more memory instead
+	regexpString := fmt.Sprintf("^(?:[a-zA-Z0-9.!#$%%&'*+\\=?^_`{|}~-]+)"+
+		"@([a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?\\.%s)$", strings.ToLower(domainName))
+	compiledRegexp, err := regexp.Compile(regexpString)
+	if err != nil {
+		return nil, err
+	}
+
+	// making rid of using deprecated package, using bufio instead
+	scanner := bufio.NewScanner(reader)
+
+	// replacing reading of the whole file with buffered line reader
+	for scanner.Scan() {
+		// need to convert bytes to lower exactly here, to let correctly pass through regex tree, despite the domain case
+		if err = json.Unmarshal(toLowerBytes(scanner.Bytes()), user); err != nil {
+			continue
 		}
 
-		if matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
-			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+		found = compiledRegexp.FindSubmatch(user.Email)
+		if len(found) < 2 {
+			continue
+		}
+
+		result[string(found[1])]++
+	}
+
+	return result, nil
+}
+
+// toLowerBytes converts ASCII bytes to lower case.
+func toLowerBytes(input []byte) []byte {
+	for index, curByte := range input {
+		if curByte >= 'A' && curByte <= 'Z' {
+			input[index] = curByte + 'a' - 'A'
 		}
 	}
-	return result, nil
+	return input
 }