--- a/hw10_program_optimization/stats_easyjson_remaked.go
+++ b/hw10_program_optimization/stats_easyjson_remaked.go
@@ -1,0 +1,62 @@
+package hw10programoptimization
+
+import (
+	"encoding/json"
+
+	"github.com/mailru/easyjson"
+	"github.com/mailru/easyjson/jlexer"
+	"github.com/mailru/easyjson/jwriter"
+)
+
+// suppress unused package warning.
+var (
+	_ *json.RawMessage
+	_ *jlexer.Lexer
+	_ *jwriter.Writer
+	_ easyjson.Marshaler
+)
+
+// easyJSONDecode method and file was generated and then manually edited to read bytes.
+func easyJSONDecode(in *jlexer.Lexer, out *User) {
+	isTopLevel := in.IsStart() //nolint
+	if in.IsNull() {
+		if isTopLevel {
+			in.Consumed()
+		}
+		in.Skip()
+		return
+	}
+	in.Delim('{')
+	for !in.IsDelim('}') {
+		key := in.UnsafeFieldName(false)
+		in.WantColon()
+		if in.IsNull() {
+			in.Skip()
+			in.WantComma()
+			continue
+		}
+		switch key {
+		case "email":
+			out.Email = in.UnsafeBytes()
+		default:
+			in.SkipRecursive()
+		}
+		in.WantComma()
+	}
+	in.Delim('}')
+	if isTopLevel {
+		in.Consumed()
+	}
+}
+
+// UnmarshalJSON supports json.Unmarshaler interface.
+func (v *User) UnmarshalJSON(data []byte) error {
+	r := jlexer.Lexer{Data: data}
+	easyJSONDecode(&r, v)
+	return r.Error()
+}
+
+// UnmarshalEasyJSON supports easyjson.Unmarshaler interface.
+func (v *User) UnmarshalEasyJSON(l *jlexer.Lexer) {
+	easyJSONDecode(l, v)
+}