--- a/hw10_program_optimization/stats_test.go
+++ b/hw10_program_optimization/stats_test.go
@@ -16,6 +16,18 @@ func TestGetDomainStat(t *testing.T) {
 {"Id":4,"Name":"Gregory Reid","Username":"tButler","Email":"5Moore@Teklist.net","Phone":"520-04-16","Password":"r639qLNu","Address":"Sunfield Park 20"}
 {"Id":5,"Name":"Janice Rose","Username":"KeithHart","Email":"nulla@Linktype.com","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}`
 
+	partiallyNonJsonData := `"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":"aliquid_qui_ea@Browsedrive.gov","Phone":"6-866-899-36-79","Password":"InAQJvsq","Address":"Blackbird Place 25"}
+{"Id":2,"Name":"Jesse Vasquez","Username":"qRichardson","Email":"mLynch@broWsecat.com","Phone":"9-373-949-64-00","Password":"SiZLeNSGn","Address":"Fulton Hill 80"}
+{"Id":3,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@Browsecat.com","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
+non json string - forget it
+{"Id":4,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@Browsecat.com",invalid json here
+{"Id":5,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@FaKeD.CoM","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
+one more non json string
+{"Id":6,"Name":"Clarence Olson","Username":"RachelAdams","Email":"invalid-mail.cOm","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
+{"Id":7,"Name":"Clarence Olson","Username":"RachelAdams","Email":"invalidmail@.com","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
+{"Id":8,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@FaKeD.cOm","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
+`
+
 	t.Run("find 'com'", func(t *testing.T) {
 		result, err := GetDomainStat(bytes.NewBufferString(data), "com")
 		require.NoError(t, err)
@@ -35,5 +47,14 @@ func TestGetDomainStat(t *testing.T) {
 		result, err := GetDomainStat(bytes.NewBufferString(data), "unknown")
 		require.NoError(t, err)
 		require.Equal(t, DomainStat{}, result)
+	})
+
+	t.Run("read through invalid records", func(t *testing.T) {
+		result, err := GetDomainStat(bytes.NewBufferString(partiallyNonJsonData), "com")
+		require.NoError(t, err)
+		require.Equal(t, DomainStat{
+			"browsecat.com": 2,
+			"faked.com":     2,
+		}, result)
 	})
 }