## Домашнее задание №10 «Оптимизация программы»

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [ ] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [ ] Я запустил `go mod tidy`.
- [ ] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [ ] Я написал бенчмарк и приложил результаты benchstat:

**Исходный бенчмарк:**
```
goos: windows
goarch: amd64
pkg: otus_go_hw/hw10_program_optimization
cpu: Intel(R) Core(TM) i7-8650U CPU @ 1.90GHz
BenchmarkGetDomainStat-8   	     111	   9146031 ns/op	 2904480 B/op	   28345 allocs/op
```
**После оптимизации:**
```
goos: windows
goarch: amd64
pkg: otus_go_hw/hw10_program_optimization
cpu: Intel(R) Core(TM) i7-8650U CPU @ 1.90GHz
BenchmarkGetDomainStat-8   	   25448	     44077 ns/op	   60214 B/op	     280 allocs/op
```
**benchcmp:**
```
benchmark                    old ns/op     new ns/op     delta
BenchmarkGetDomainStat-8     9146031       44077         -99.52%

benchmark                    old allocs     new allocs     delta
BenchmarkGetDomainStat-8     28345          280            -99.01%

benchmark                    old bytes     new bytes     delta
BenchmarkGetDomainStat-8     2904480       60214         -97.93%
```
**benchstat:**
```
name             old time/op    new time/op    delta
GetDomainStat-8    9.15ms ± 0%    0.04ms ± 0%   ~     (p=1.000 n=1+1)

name             old alloc/op   new alloc/op   delta
GetDomainStat-8    2.90MB ± 0%    0.06MB ± 0%   ~     (p=1.000 n=1+1)

name             old allocs/op  new allocs/op  delta
GetDomainStat-8     28.3k ± 0%      0.3k ± 0%   ~     (p=1.000 n=1+1)
```

### Критерии оценки
- [ ] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [ ] Добавлены юнит-тесты - до 3 баллов
- [ ] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов